decided to create easier to understand api to use as seed / portfolio. going with school > class > student type api seed project
